Split test files; move checkpoint/fixture-gated tests to IntegrationTesting
Two test files in `Tests/MLXLMTests/` mixed synthetic-mock cases (no
fixture or checkpoint dependency, CI-runnable) with checkpoint or
fixture-gated cases. Splitting keeps the synthetic half in `MLXLMTests`
so they continue to run on CI, and moves the integration half into the
IntegrationTesting Xcode project alongside the four files moved in the
previous commit.

`MTPDrafterModelFactoryTests.swift`:

- Stays in `Tests/MLXLMTests/`: `testGemma4AssistantRegistrationRegistersType`,
  `testMTPDrafterTypeRegistryUnknownModelTypeThrows`,
  `testMTPDrafterRegistryContainsBothReferenceCheckpoints`,
  `testMTPDrafterRegistrySharedStaticAccessors`.
- New `MTPDrafterModelFactoryIntegrationTests.swift` in
  `IntegrationTesting/IntegrationTestingTests/`:
  `testMTPDrafterFactoryLoadFromDirectoryWhenCheckpointPresent` plus
  the `NoOpTokenizerLoader` helper it consumes.

`Gemma4AssistantDraftModelTests.swift`:

- Stays in `Tests/MLXLMTests/`: 4 synthetic tests (config JSON decode,
  `sanitize` behavior under tied/untied embeddings, instantiation
  shape) plus the `syntheticConfig(...)` helper. Drops the
  `@_spi(Testing) import MLXVLM` since `forwardHidden` (the only
  SPI-Testing symbol) is no longer referenced here.
- New `Gemma4AssistantDraftModelIntegrationTests.swift` in
  `IntegrationTesting/IntegrationTestingTests/`: the 3 checkpoint or
  fixture-gated tests
  (`testGemma4AssistantConfigurationDecodesRealCheckpoint`,
  `testRung1WeightsLoadFrom31BCheckpoint`,
  `testRung2And3ForwardMatchesFixture31BCase01`), with the Rung 2/3
  forward parity test rewired to fetch fixtures via `downloadDataset(...)`
  from the pinned HF dataset and the `Gemma4AssistantDraftModel.
  forwardHidden(...)` SPI re-imported under `@_spi(Testing)`.

Verified locally: `swift build --build-tests` compiles `MLXLMTests` and
all dependent targets clean; `xcodebuild build-for-testing -scheme
IntegrationTesting` compiles all moved + new integration test files.
